Update README.md
Explaining why I added my work to a RMD file instead of a MD file. In the future I might try to have my RMD file create a "md_document" with the variant as "markdown_github" if that's easier. I could specify this in the metadata or the YAML of RMD.